feat(pagenotfound): migration typescript
