--- src/pages/NotFound/__tests__/NotFound.spec.js
@@ -1,29 +1,0 @@
-import { render } from '@testing-library/react';
-import { StaticRouter } from 'react-router-dom/server';
-import { emptyFunction } from 'shared/testsUtils';
-import { defineFeature, loadFeature } from 'jest-cucumber';
-import NotFound from '../NotFound';
-
-const wrapper = ({ children }) => <StaticRouter context={{}}>{children}</StaticRouter>;
-
-const feature = loadFeature('src/pages/NotFound/NotFound.feature');
-
-defineFeature(feature, test => {
-  test('Chargement de la page 404', ({ given, when, then }) => {
-    let $;
-
-    given('Je suis un utilisateur accédant à une page inconnue', emptyFunction);
-
-    when('Je suis redirigé sur cette page', () => {
-      $ = render(<NotFound />, { wrapper });
-    });
-
-    then('La page m\'affiche un titre "404notfound"', () => {
-      expect($.getByText('404 Page Not Found')).toBeDefined();
-    });
-
-    then('La page m\'affiche le contenu suivant "The page you are looking for is not here!"', () => {
-      expect($.getByText('The page you are looking for is not here!')).toBeDefined();
-    });
-  });
-});
