--- src/pages/NotFound/__tests__/NotFound.test.js
@@ -1,12 +1,0 @@
-import { StaticRouter } from 'react-router-dom/server';
-import { render } from '@testing-library/react';
-import NotFound from '../NotFound';
-
-const wrapper = ({ children }) => <StaticRouter context={{}}>{children}</StaticRouter>;
-
-describe('<NotFound page/>', () => {
-  it('Renders NotFound page component without crashing', () => {
-    const { asFragment } = render(<NotFound />, { wrapper });
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
