+++ src/pages/NotFound/__tests__/NotFound.test.tsx
@@ -1,0 +1,9 @@
+import { renderWithWrapperStaticRouter } from 'shared/testsUtils';
+import NotFound from '../NotFound';
+
+describe('<NotFound page/>', () => {
+  it('Renders NotFound page component without crashing', () => {
+    const { asFragment } = renderWithWrapperStaticRouter(<NotFound />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
